chore: move schema code into separate dir
