+++ src/schema/index.ts
@@ -1,0 +1,2 @@
+export * from './cdk-schema';
+export * from './jsii2schema';
